backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiohttp==3.9.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.24.3
scipy==1.11.4
faster-whisper==0.9.0
soundfile==0.12.1
PyPDF2==3.0.1
